system-tests/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
docker==7.0.0
paho-mqtt==1.6.1
system-tests/src/resources/devispare_python-1.3.1-py3-none-any.whl
littlefs-python==0.9.1
allure-pytest==2.13.2
pyhamcrest==2.1.0
cryptography==42.0.5
